+++ projects/js-packages/ai-client/src/audio-transcription/index.ts
@@ -7,10 +7,6 @@ import debugFactory from 'debug';
  */
 import apiFetch from '../api-fetch/index.js';
 import requestJwt from '../jwt/index.js';
-/**
- * Types
- */
-import { CancelablePromise } from '../types.js';
 
 const debug = debugFactory( 'jetpack-ai-client:audio-transcription' );
 
@@ -29,13 +25,14 @@ type AudioTranscriptionResponse = {
  *
  * @param {Blob} audio - The audio to be transcribed, from a recording or from a file.
  * @param {string} feature - The feature name that is calling the transcription.
+ * @param {AbortSignal} requestAbortSignal - The signal to abort the request.
  * @returns {Promise<string>} - The promise of a string containing the transcribed audio.
  */
 export default async function transcribeAudio(
 	audio: Blob,
-	feature?: string
-	// @ts-expect-error Promises are not cancelable by default
-): CancelablePromise< string > {
+	feature?: string,
+	requestAbortSignal?: AbortSignal
+): Promise< string > {
 	debug( 'Transcribing audio: %o. Feature: %o', audio, feature );
 
 	// Get a token to use the transcription service
@@ -63,6 +60,7 @@ export default async function transcribeAudio(
 			method: 'POST',
 			body: formData,
 			headers,
+			signal: requestAbortSignal ?? undefined,
 		} );
 
 		debug( 'Transcription response: %o', response );