+++ projects/packages/my-jetpack/_inc/components/redeem-token-screen/index.jsx
@@ -2,7 +2,8 @@ import { useConnection } from '@automattic/jetpack-connection';
 import { GoldenTokenModal } from '@automattic/jetpack-licensing';
 import { __ } from '@wordpress/i18n';
 import React from 'react';
-import usePurchases from '../../hooks/use-purchases';
+import { REST_API_SITE_PURCHASES_ENDPOINT } from '../../data/constants';
+import useSimpleQuery from '../../data/use-simple-query';
 import { includesLifetimePurchase } from '../../utils/is-lifetime-purchase';
 
 /**
@@ -17,10 +18,13 @@ export default function RedeemTokenScreen() {
 		userConnectionData?.currentUser?.wpcomUser?.display_name ||
 		userConnectionData?.currentUser?.wpcomUser?.login ||
 		userConnectionData?.currentUser?.username;
-	const { isFetchingPurchases, purchases } = usePurchases();
+	const { isLoading, data: purchases } = useSimpleQuery( 'purchases', {
+		path: REST_API_SITE_PURCHASES_ENDPOINT,
+	} );
+
 	const tokenRedeemed = includesLifetimePurchase( purchases );
 
-	if ( isFetchingPurchases ) {
+	if ( isLoading ) {
 		return <>{ __( 'Checking gold status…', 'jetpack-my-jetpack' ) }</>;
 	}