+++ projects/plugins/boost/app/assets/src/js/features/page-cache/meta/meta.tsx
@@ -1,4 +1,4 @@
-import { Button, Notice } from '@automattic/jetpack-components';
+import { Button, IconTooltip, Notice } from '@automattic/jetpack-components';
 import { createInterpolateElement } from '@wordpress/element';
 import { __, _n, sprintf } from '@wordpress/i18n';
 import ChevronDown from '$svg/chevron-down';
@@ -57,14 +57,23 @@ const Meta = () => {
 		);
 	};
 
+	const loggingEnabledMessage = __( 'Logging enabled.', 'jetpack-boost' );
+	const loggingDisabledMessage = __( 'Logging disabled.', 'jetpack-boost' );
+
 	return (
 		<div className={ styles.wrapper }>
-			<MutationNotice { ...mutateBypassPatterns } />
+			<MutationNotice { ...mutateBypassPatterns } mutationId="update-bypass-patterns" />
+			<MutationNotice
+				{ ...mutateLogging }
+				successMessage={ logging ? loggingEnabledMessage : loggingDisabledMessage }
+				mutationId="update-logging"
+			/>
 			<MutationNotice
 				{ ...runClearPageCacheAction }
 				savingMessage={ __( 'Clearing cache…', 'jetpack-boost' ) }
 				errorMessage={ __( 'Unable to clear cache.', 'jetpack-boost' ) }
 				successMessage={ clearedCacheMessage || __( 'Cache cleared.', 'jetpack-boost' ) }
+				mutationId="clear-page-cache"
 			/>
 			<div className={ styles.head }>
 				<div className={ styles.summary }>{ getSummary() }</div>
@@ -104,7 +113,7 @@ const Meta = () => {
 						/>
 						<div className={ styles.section }>
 							<div className={ styles.title }>{ __( 'Logging', 'jetpack-boost' ) }</div>
-							<label htmlFor="cache-logging">
+							<label htmlFor="cache-logging" className={ styles[ 'logging-toggle' ] }>
 								<input
 									type="checkbox"
 									id="cache-logging"
@@ -112,13 +121,13 @@ const Meta = () => {
 									onChange={ event => mutateLogging.mutate( event.target.checked ) }
 								/>{ ' ' }
 								{ __( 'Activate logging to track all your cache events.', 'jetpack-boost' ) }
-								{ logging && (
-									<>
-										{ ' ' }
-										<Link to="/cache-debug-log">{ __( 'See Logs', 'jetpack-boost' ) }</Link>
-									</>
-								) }
 							</label>
+							{ logging && (
+								<Link className={ styles[ 'see-logs-link' ] } to="/cache-debug-log">
+									{ __( 'See Logs', 'jetpack-boost' ) }
+								</Link>
+							) }
+							<div className={ styles.clearfix } />
 						</div>
 					</>
 				</div>
@@ -148,6 +157,27 @@ const BypassPatterns = ( {
 	// @todo - add proper link.
 	const exclusionsLink = 'https://jetpack.com';
 
+	const validateInputValue = ( value: string ) => {
+		setInputValue( value );
+		setInputInvalid( ! validatePatterns( value ) );
+	};
+
+	const validatePatterns = ( value: string ) => {
+		const lines = value
+			.split( '\n' )
+			.map( line => line.trim() )
+			.filter( line => line.trim() !== '' );
+
+		// check if it's a valid regex
+		try {
+			lines.forEach( line => new RegExp( line ) );
+		} catch ( e ) {
+			return false;
+		}
+
+		return true;
+	};
+
 	useEffect( () => {
 		setInputValue( patterns );
 	}, [ patterns ] );
@@ -173,13 +203,13 @@ const BypassPatterns = ( {
 			<textarea
 				value={ inputValue }
 				rows={ 3 }
-				onChange={ e => setInputValue( e.target.value ) }
+				onChange={ e => validateInputValue( e.target.value ) }
 				id="jb-cache-exceptions"
 			/>
 			<p className={ classNames( styles.description, styles[ 'error-message' ] ) }>
 				{ __( 'Error: Invalid format', 'jetpack-boost' ) }
 			</p>
-			<p className={ styles.description }>
+			<div className={ styles.description }>
 				{ __(
 					'Use (.*) to address multiple URLs under a given path. Be sure each URL path is in its own line.',
 					'jetpack-boost'
@@ -188,13 +218,12 @@ const BypassPatterns = ( {
 				{ createInterpolateElement(
 					__( '<help>See an example</help> or <link>learn more</link>.', 'jetpack-boost' ),
 					{
-						// eslint-disable-next-line jsx-a11y/anchor-has-content, jsx-a11y/anchor-is-valid
-						help: <a href="#" target="_blank" rel="noreferrer" />,
+						help: <BypassPatternsExample />, // children are passed after the interpolation.
 						// eslint-disable-next-line jsx-a11y/anchor-has-content
 						link: <a href={ exclusionsLink } target="_blank" rel="noreferrer" />,
 					}
 				) }
-			</p>
+			</div>
 			{ showNotice && (
 				<Notice
 					level="error"
@@ -204,9 +233,54 @@ const BypassPatterns = ( {
 					{ __( 'An error occurred while saving changes. Please, try again.', 'jetpack-boost' ) }
 				</Notice>
 			) }
-			<Button disabled={ patterns === inputValue } onClick={ save } className={ styles.button }>
+			<Button
+				disabled={ patterns === inputValue || inputInvalid }
+				onClick={ save }
+				className={ styles.button }
+			>
 				{ __( 'Save', 'jetpack-boost' ) }
 			</Button>
+		</div>
+	);
+};
+
+type BypassPatternsExampleProps = {
+	children?: React.ReactNode;
+};
+
+const BypassPatternsExample = ( { children }: BypassPatternsExampleProps ) => {
+	const [ show, setShow ] = useState( false );
+
+	return (
+		<div className={ styles[ 'example-wrapper' ] }>
+			{ /* eslint-disable-next-line jsx-a11y/anchor-is-valid */ }
+			<a
+				href="#"
+				className={ styles[ 'example-button' ] }
+				onClick={ e => {
+					e.preventDefault();
+					setShow( ! show );
+				} }
+			>
+				{ children }
+			</a>
+			<div className={ styles[ 'tooltip-wrapper' ] }>
+				<IconTooltip
+					placement="bottom-start"
+					popoverAnchorStyle="wrapper"
+					forceShow={ show }
+					offset={ -10 }
+					className={ styles.tooltip }
+				>
+					<strong>{ __( 'Example:', 'jetpack-boost' ) }</strong>
+					<br />
+					checkout
+					<br />
+					gallery/.*
+					<br />
+					specific-page
+				</IconTooltip>
+			</div>
 		</div>
 	);
 };