+++ projects/packages/publicize/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.42.3] - 2024-02-26
+### Fixed
+- Deprecate the sharing_menu method of the Publicize_UI class. [#35810]
+
 ## [0.42.2] - 2024-02-14
 ### Fixed
 - Fixed an issue where on old sites og:image is an array that causes issues [#35688]
@@ -480,6 +484,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies.
 - Update package.json metadata.
 
+[0.42.3]: https://github.com/Automattic/jetpack-publicize/compare/v0.42.2...v0.42.3
 [0.42.2]: https://github.com/Automattic/jetpack-publicize/compare/v0.42.1...v0.42.2
 [0.42.1]: https://github.com/Automattic/jetpack-publicize/compare/v0.42.0...v0.42.1
 [0.42.0]: https://github.com/Automattic/jetpack-publicize/compare/v0.41.0...v0.42.0